--- a/sparkcc.py
+++ b/sparkcc.py
@@ -11,11 +11,11 @@ import boto3
 import botocore
 import requests
 
-from warcio.archiveiterator import ArchiveIterator
-from warcio.recordloader import ArchiveLoadFailed
-
 from pyspark.sql import SparkSession
 from pyspark.sql.types import StructType, StructField, StringType, LongType
+
+from warcio.archiveiterator import ArchiveIterator
+from warcio.recordloader import ArchiveLoadFailed
 
 
 LOGGING_FORMAT = '%(asctime)s %(levelname)s %(name)s: %(message)s'
@@ -24,6 +24,7 @@ LOGGING_FORMAT = '%(asctime)s %(levelname)s %(name)s: %(message)s'
 class CCSparkJob(object):
     """
     A simple Spark job definition to process Common Crawl data
+    (WARC/WAT/WET files using Spark and warcio)
     """
 
     name = 'CCSparkJob'
@@ -146,7 +147,6 @@ class CCSparkJob(object):
         logging.getLogger(self.name).setLevel(level)
         if session:
             session.sparkContext.setLogLevel(level)
-
 
     def init_accumulators(self, session):
         """Register and initialize counters (aka. accumulators).
@@ -348,18 +348,23 @@ class CCSparkJob(object):
             if not stream:
                 continue
 
-            no_parse = (not self.warc_parse_http_header)
-            try:
-                archive_iterator = ArchiveIterator(stream,
-                                                   no_record_parse=no_parse, arc2warc=True)
-                for res in self.iterate_records(uri, archive_iterator):
-                    yield res
-            except ArchiveLoadFailed as exception:
-                self.warc_input_failed.add(1)
-                self.get_logger().error(
-                    'Invalid WARC: {} - {}'.format(uri, exception))
-            finally:
-                stream.close()
+            for res in self.process_warc(uri, stream):
+                yield res
+
+            stream.close()
+
+    def process_warc(self, uri, stream):
+        """Parse a WARC (or WAT/WET file) using warcio,
+        call iterate_records() to process the WARC records"""
+        try:
+            rec_iter = ArchiveIterator(stream,
+                                       no_record_parse=(not self.warc_parse_http_header),
+                                       arc2warc=True)
+            for res in self.iterate_records(uri, rec_iter):
+                yield res
+        except ArchiveLoadFailed as exception:
+            self.warc_input_failed.add(1)
+            self.get_logger().error('Invalid WARC: {} - {}'.format(uri, exception))
 
     def process_record(self, record):
         """Process a single WARC/WAT/WET record"""