--- a/sparkcc.py
+++ b/sparkcc.py
@@ -15,7 +15,7 @@ from pyspark.sql import SparkSession
 from pyspark.sql.types import StructType, StructField, StringType, LongType
 
 from warcio.archiveiterator import ArchiveIterator
-from warcio.recordloader import ArchiveLoadFailed
+from warcio.recordloader import ArchiveLoadFailed, ArcWarcRecord
 
 
 LOGGING_FORMAT = '%(asctime)s %(levelname)s %(name)s: %(message)s'
@@ -385,19 +385,36 @@ class CCSparkJob(object):
             #  warc_record_length = archive_iterator.get_record_length()
 
     @staticmethod
-    def is_wet_text_record(record):
+    def get_payload_stream(record: ArcWarcRecord):
+        return record.content_stream()
+
+    @staticmethod
+    def get_warc_header(record: ArcWarcRecord, header: str):
+        return record.rec_headers.get_header(header)
+
+    @staticmethod
+    def get_http_headers(record: ArcWarcRecord):
+        return record.http_headers.headers
+
+    @staticmethod
+    def is_response_record(record: ArcWarcRecord):
+        """Return true if WARC record is a WARC response record"""
+        return record.rec_type == 'response'
+
+    @staticmethod
+    def is_wet_text_record(record: ArcWarcRecord):
         """Return true if WARC record is a WET text/plain record"""
         return (record.rec_type == 'conversion' and
                 record.content_type == 'text/plain')
 
     @staticmethod
-    def is_wat_json_record(record):
+    def is_wat_json_record(record: ArcWarcRecord):
         """Return true if WARC record is a WAT record"""
         return (record.rec_type == 'metadata' and
                 record.content_type == 'application/json')
 
     @staticmethod
-    def is_html(record):
+    def is_html(record: ArcWarcRecord):
         """Return true if (detected) MIME type of a record is HTML"""
         html_types = ['text/html', 'application/xhtml+xml']
         if (('WARC-Identified-Payload-Type' in record.rec_headers) and